requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
future==0.18.2
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipfshttpclient==0.8.0a2
jsonschema==4.6.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
pathspec==0.9.0
pefile==2022.5.30
platformdirs==2.5.2
protobuf==3.20.1
pycryptodome==3.14.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pyrsistent==0.18.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
rlp==2.0.1
six==1.16.0
tomli==2.0.1
toolz==0.11.2
typed-ast==1.5.4
typing_extensions==4.2.0
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
websockets==9.1
yarl==1.7.2
zipp==3.8.0
